// src/pages/metrics-pharmacy/model/pharmacyMetricsTypes.ts
export type PharmacyMetrics = {
  id: number;
  pharmacy_id: number;
  score: string;
  total_orders: number;
  completed_orders: number;
  completion_rate: string;
  avg_items_per_order: string;
  total_items: number;
  last_order_at: string;
  recency_score: string;
  calculated_at: string;
  created_at: string;
  updated_at: string;
};